google_ads_api.py
from google.ads.googleads.client import GoogleAdsClient
from utils import log_info

def initialize_client(config):
    """
    Autentica e inicializa o cliente do Google Ads.
    """
    client = GoogleAdsClient.load_from_dict(config['google_ads'])
    log_info("Cliente do Google Ads inicializado")
    return client

def create_campaign(client, customer_id, campaign_config):
    """
    Cria campanha com configurações fornecidas.
    """
    # Implementação dos detalhes da criação de campanha...
    log_info(f"Campanha criada para cliente {customer_id}")

