requirements.txt
google-ads>=19.0.0
pandas>=2.0.0
openai>=0.27.0
PyYAML>=6.0

